code refactoring and making some components smaller to prepare for testing
