+++ src/components/customers/__tests__/CustomerPaymentMethods.test.tsx
@@ -28,7 +28,7 @@ jest.mock('~/core/utils/copyToClipboard', () => ({
   copyToClipboard: jest.fn(),
 }))
 
-jest.mock('~/components/customers/PaymentMethodsList', () => ({
+jest.mock('~/components/customers/paymentMethodsList', () => ({
   PaymentMethodsList: () => <div>Payment Methods List</div>,
 }))
