+++ src/components/customers/paymentMethodsList/index.tsx
@@ -1,35 +1,17 @@
 import { gql } from '@apollo/client'
-import { Icon, IconName, Typography } from 'lago-design-system'
+import { useCallback } from 'react'
 
-import { Status, StatusType } from '~/components/designSystem'
-import { Chip } from '~/components/designSystem/Chip'
-import { ActionItem, Table } from '~/components/designSystem/Table'
-import { PaymentProviderChip } from '~/components/PaymentProviderChip'
-import { addToast } from '~/core/apolloClient'
-import { copyToClipboard } from '~/core/utils/copyToClipboard'
+import { Table } from '~/components/designSystem/Table'
 import {
   DestroyPaymentMethodInput,
-  PaymentMethodsQuery,
-  ProviderPaymentMethodsEnum,
   SetAsDefaultInput,
   useDestroyPaymentMethodMutation,
   usePaymentMethodsQuery,
   useSetPaymentMethodAsDefaultMutation,
 } from '~/generated/graphql'
-import { TranslateFunc, useInternationalization } from '~/hooks/core/useInternationalization'
+import { useInternationalization } from '~/hooks/core/useInternationalization'
 
-const OBFUSCATED_LAST4_PREFIX = '••••'
-
-const PaymentProviderMethodTranslationsLookup: Record<ProviderPaymentMethodsEnum, IconName> = {
-  [ProviderPaymentMethodsEnum.BacsDebit]: 'bank',
-  [ProviderPaymentMethodsEnum.Card]: 'card',
-  [ProviderPaymentMethodsEnum.Link]: 'bank',
-  [ProviderPaymentMethodsEnum.SepaDebit]: 'bank',
-  [ProviderPaymentMethodsEnum.UsBankAccount]: 'bank',
-  [ProviderPaymentMethodsEnum.Boleto]: 'bank',
-  [ProviderPaymentMethodsEnum.Crypto]: 'bank',
-  [ProviderPaymentMethodsEnum.CustomerBalance]: 'bank',
-}
+import { usePaymentMethodsTableColumns } from './usePaymentMethodsTableColumns'
 
 gql`
   query PaymentMethods($externalCustomerId: ID!) {
@@ -68,61 +50,14 @@ interface Props {
   externalCustomerId: string
 }
 
-type PaymentMethodItem = PaymentMethodsQuery['paymentMethods']['collection'][number]
-
-const generateActionColumn = ({
-  translate,
-  setPaymentMethodAsDefault,
-  destroyPaymentMethod,
-}: {
-  translate: TranslateFunc
-  setPaymentMethodAsDefault: (input: SetAsDefaultInput) => Promise<void>
-  destroyPaymentMethod: (input: DestroyPaymentMethodInput) => Promise<void>
-}): ActionItem<PaymentMethodItem>[] => {
-  const setPaymentMethodAsDefaultAction: ActionItem<PaymentMethodItem> = {
-    startIcon: 'star-filled',
-    title: translate('text_1728574726495n9jdse2hnrf'),
-    onAction: async (paymentMethod) => {
-      await setPaymentMethodAsDefault({ id: paymentMethod.id })
-
-      addToast({
-        severity: 'success',
-        translateKey: translate('text_1762437511802do7br4qjc5p'),
-      })
-    },
-  }
-
-  const copyPaymentMethodAction: ActionItem<PaymentMethodItem> = {
-    startIcon: 'duplicate',
-    title: translate('text_1762437511802p8qb8qpfy9a'),
-    onAction: (paymentMethod) => {
-      copyToClipboard(paymentMethod.id)
-
-      addToast({
-        severity: 'info',
-        translateKey: translate('text_176243751180284tefl8wgpw'),
-      })
-    },
-  }
-
-  const deletePaymentMethodAction: ActionItem<PaymentMethodItem> = {
-    startIcon: 'trash',
-    title: translate('text_1762437511802sg9jrl46lkb'),
-    onAction: async (paymentMethod) => {
-      await destroyPaymentMethod({ id: paymentMethod.id })
-
-      addToast({
-        severity: 'success',
-        translateKey: translate('text_1762437511802g5ysxig14q5'),
-      })
-    },
-  }
-
-  return [setPaymentMethodAsDefaultAction, copyPaymentMethodAction, deletePaymentMethodAction]
-}
-
 export const PaymentMethodsList = ({ externalCustomerId }: Props) => {
   const { translate } = useInternationalization()
+
+  const [setPaymentMethodAsDefaultMutation, { error: errorSetAsDefault }] =
+    useSetPaymentMethodAsDefaultMutation()
+
+  const [destroyPaymentMethodMutation, { error: errorDestroyPaymentMethod }] =
+    useDestroyPaymentMethodMutation()
 
   const {
     data,
@@ -133,26 +68,29 @@ export const PaymentMethodsList = ({ externalCustomerId }: Props) => {
     variables: { externalCustomerId },
   })
 
-  const [setPaymentMethodAsDefaultMutation, { error: hasErrorSetPaymentMethodAsDefault }] =
-    useSetPaymentMethodAsDefaultMutation()
-  const [destroyPaymentMethodMutation, { error: hasErrorDestroyPaymentMethod }] =
-    useDestroyPaymentMethodMutation()
+  const setPaymentMethodAsDefault = useCallback(
+    async (input: SetAsDefaultInput): Promise<void> => {
+      await setPaymentMethodAsDefaultMutation({ variables: { input } })
+      refetch()
+    },
+    [setPaymentMethodAsDefaultMutation, refetch],
+  )
 
-  const setPaymentMethodAsDefault = async (input: SetAsDefaultInput): Promise<void> => {
-    await setPaymentMethodAsDefaultMutation({ variables: { input } })
-    refetch()
-  }
+  const destroyPaymentMethod = useCallback(
+    async (input: DestroyPaymentMethodInput): Promise<void> => {
+      await destroyPaymentMethodMutation({ variables: { input } })
+      refetch()
+    },
+    [destroyPaymentMethodMutation, refetch],
+  )
 
-  const destroyPaymentMethod = async (input: DestroyPaymentMethodInput): Promise<void> => {
-    await destroyPaymentMethodMutation({ variables: { input } })
-    refetch()
-  }
+  const { columns, actionColumn } = usePaymentMethodsTableColumns({
+    setPaymentMethodAsDefault,
+    destroyPaymentMethod,
+  })
 
-  const paymentMethods: PaymentMethodsQuery['paymentMethods']['collection'] =
-    data?.paymentMethods?.collection || []
-
-  const hasError =
-    hasErrorPaymentMethods || hasErrorSetPaymentMethodAsDefault || hasErrorDestroyPaymentMethod
+  const paymentMethods = data?.paymentMethods?.collection || []
+  const hasError = hasErrorPaymentMethods || errorSetAsDefault || errorDestroyPaymentMethod
 
   return (
     <Table
@@ -167,120 +105,10 @@ export const PaymentMethodsList = ({ externalCustomerId }: Props) => {
         },
       }}
       actionColumnTooltip={() => translate('text_634687079be251fdb438338f')}
-      actionColumn={() =>
-        generateActionColumn({ translate, setPaymentMethodAsDefault, destroyPaymentMethod })
-      }
-      columns={[
-        {
-          key: 'id',
-          title: translate('text_1762437511802dynl0tx20xe'),
-          maxSpace: true,
-          content: (item) => {
-            const paymentDetailType = (item.details?.type || '').toLowerCase()
-            const iconName =
-              paymentDetailType && paymentDetailType in PaymentProviderMethodTranslationsLookup
-                ? PaymentProviderMethodTranslationsLookup[
-                    paymentDetailType as ProviderPaymentMethodsEnum
-                  ]
-                : 'bank'
-
-            return (
-              <div className="flex items-center gap-3">
-                {/* ICON */}
-                <div className="flex size-10 items-center justify-center rounded-xl bg-grey-100">
-                  <Icon name={iconName} color="black" />
-                </div>
-
-                <div className="flex flex-1 flex-col">
-                  <div className="flex items-center gap-1">
-                    {/* PAYMENT METHOD DETAILS */}
-                    {item.details?.type && (
-                      <Typography variant="body" className="capitalize text-grey-700">
-                        {item.details.type}
-                      </Typography>
-                    )}
-                    {item.details?.type && item.details?.brand && (
-                      <Typography variant="body" className="text-grey-700">
-                        {' - '}
-                      </Typography>
-                    )}
-                    {item.details?.brand && (
-                      <Typography variant="body" className="capitalize text-grey-700">
-                        {item.details.brand}
-                      </Typography>
-                    )}
-                    {item.details?.last4 && (
-                      <Typography variant="body" className="text-grey-700">
-                        {OBFUSCATED_LAST4_PREFIX} {item.details.last4}
-                      </Typography>
-                    )}
-
-                    {/* EXPIRATION DATE */}
-                    {item.details?.expirationMonth && item.details?.expirationYear && (
-                      <Chip
-                        label={`${translate('text_1762437511802zhw5mx0iamd')} ${item.details.expirationMonth}/${item.details.expirationYear}`}
-                        type="primary"
-                        color="grey700"
-                        variant="caption"
-                        size="small"
-                        className="ml-2"
-                      />
-                    )}
-
-                    {/* DEFAULT BADGE */}
-                    {item.isDefault && (
-                      <Chip
-                        label={translate('text_17440321235444hcxi31f8j6')}
-                        type="secondary"
-                        variant="caption"
-                        color="info600"
-                        size="small"
-                        className="ml-2 bg-purple-100"
-                      />
-                    )}
-                  </div>
-
-                  {/* PSP INFO */}
-                  <div className="flex items-center gap-1">
-                    {item.paymentProviderType && (
-                      <PaymentProviderChip
-                        paymentProvider={item.paymentProviderType}
-                        className="text-xs"
-                        textVariant="caption"
-                        textColor="grey500"
-                      />
-                    )}
-                    {item.paymentProviderType && item.paymentProviderCode && (
-                      <Typography variant="caption" className="text-grey-500">
-                        {' • '}
-                      </Typography>
-                    )}
-                    {item.paymentProviderCode && (
-                      <Typography variant="caption" className="text-grey-500">
-                        {item.paymentProviderCode}
-                      </Typography>
-                    )}
-                  </div>
-                </div>
-              </div>
-            )
-          },
-        },
-        {
-          key: 'id',
-          title: translate('text_63ac86d797f728a87b2f9fa7'),
-          content: () => (
-            <Typography variant="caption" color="textSecondary">
-              {/* <Status type={StatusType.success} label="active" /> */}
-              <Status type={StatusType.disabled} label="Inactive" />
-            </Typography>
-          ),
-        },
-      ]}
+      actionColumn={actionColumn}
+      columns={columns}
       isLoading={loading}
       hasError={!!hasError}
     />
   )
 }
-
-PaymentMethodsList.displayName = 'PaymentMethodsList'
