+++ translations/base.json
@@ -3508,5 +3508,6 @@
   "text_1762437511802g5ysxig14q5": "Payment method ID successfully deleted",
   "text_1762437511802dynl0tx20xe": "Payment Method",
   "text_1762437511802do7br4qjc5p": "Payment method successfully set as default",
-  "text_1762437511802zhw5mx0iamd": "Exp"
+  "text_1762437511802zhw5mx0iamd": "Exp",
+  "text_17625289719370dmo0r5s8c3": "Inactive"
 }
